+++ pynmeagps/nmeareader.py
@@ -10,7 +10,7 @@ Returns both the raw binary data (as bytes) and the parsed
 data (as a NMEAMessage object).
 
 If the 'nmeaonly' kwarg is set to 'True', the reader
-will raise a NMEAStreamerError if it encounters any non-NMEA
+will raise a NMEAParseError if it encounters any non-NMEA
 data. Otherwise, it will ignore the non-NMEA data and attempt
 to carry on.
 
@@ -30,7 +30,14 @@ from pynmeagps.nmeahelpers import (
     calc_checksum,
     isvalid_cksum,
 )
-from pynmeagps.nmeatypes_core import NMEA_HDR, VALCKSUM, VALMSGID
+from pynmeagps.nmeatypes_core import (
+    NMEA_HDR,
+    VALCKSUM,
+    VALMSGID,
+    ERR_IGNORE,
+    ERR_LOG,
+    ERR_RAISE,
+)
 
 
 class NMEAReader:
@@ -42,11 +49,13 @@ class NMEAReader:
         """Constructor.
 
         :param stream stream: input data stream (e.g. Serial or binary File)
+        :param int quitonerror: (kwarg) 0 = ignore errors,  1 = log errors and continue, 2 = (re)raise errors (1)
+        :param int errorhandler: (kwarg) error handling object or function (None)
         :param bool nmeaonly (kwarg): True = error on non-NMEA data, False = ignore non-NMEA data
         :param int validate (kwarg): bitfield validation flags - VALCKSUM (default), VALMSGID
         :param int msgmode (kwarg): 0 = GET (default), 1 = SET, 2 = POLL
         :param int bufsize: (kwarg) socket recv buffer size (4096)
-        :raises: NMEAStreamError (if mode is invalid)
+        :raises: NMEAParseError (if mode is invalid)
 
         """
 
@@ -59,9 +68,11 @@ class NMEAReader:
         validate = kwargs.get("validate", VALCKSUM)
         msgmode = kwargs.get("msgmode", 0)
         if msgmode not in (0, 1, 2):
-            raise nme.NMEAStreamError(
+            raise nme.NMEAParseError(
                 f"Invalid stream mode {msgmode} - must be 0, 1 or 2."
             )
+        self._quitonerror = kwargs.get("quitonerror", ERR_LOG)
+        self._errorhandler = kwargs.get("errorhandler", None)
         self._nmea_only = nmeaonly
         self._validate = validate
         self._mode = msgmode
@@ -82,9 +93,9 @@ class NMEAReader:
         """
 
         (raw_data, parsed_data) = self.read()
-        if raw_data is not None:
-            return (raw_data, parsed_data)
-        raise StopIteration
+        if raw_data is None and parsed_data is None:
+            raise StopIteration
+        return (raw_data, parsed_data)
 
     def read(self) -> (bytes, NMEAMessage):
         """
@@ -96,68 +107,104 @@ class NMEAReader:
 
         """
 
-        reading = True
+        parsing = True
         raw_data = None
         parsed_data = None
 
-        while reading:  # loop until end of valid NMEA message or EOF
-            byte1 = self._stream.read(1)  # read 1st byte
-            if len(byte1) < 1:  # EOF
-                break
-            if byte1 != b"\x24":  # not NMEA, discard and continue
-                continue
-            byte2 = self._stream.read(1)  # read 2nd byte to confirm protocol
-            if len(byte2) < 1:  # EOF
-                break
-            bytehdr = byte1 + byte2
-            if bytehdr in NMEA_HDR:  # it's a NMEA message
-                byten = self._stream.readline()  # NMEA protocol is CRLF terminated
-                if byten[-2:] != b"\x0d\x0a":  # EOF
-                    break
-                raw_data = bytehdr + byten
-                parsed_data = self.parse(
-                    raw_data, validate=self._validate, msgmode=self._mode
-                )
-                reading = False
-            else:  # it's not a NMEA message (UBX or something else)
-                if self._nmea_only:  # raise error and quit
-                    raise nme.NMEAStreamError(f"Unknown data header {bytehdr}.")
+        try:
+            while parsing:  # loop until end of valid NMEA message or EOF
+                byte1 = self._read_bytes(1)  # read 1st byte
+                if byte1 != b"\x24":  # not NMEA, discard and continue
+                    continue
+                byte2 = self._read_bytes(1)  # read 2nd byte to confirm protocol
+                bytehdr = byte1 + byte2
+                if bytehdr in NMEA_HDR:  # it's a NMEA message
+                    byten = self._read_line()  # NMEA protocol is CRLF terminated
+                    raw_data = bytehdr + byten
+                    parsed_data = self.parse(
+                        raw_data, validate=self._validate, msgmode=self._mode
+                    )
+                    parsing = False
+                else:  # it's not a NMEA message (UBX or something else)
+                    if self._nmea_only:  # raise error and quit
+                        raise nme.NMEAParseError(f"Unknown data header {bytehdr}.")
+
+        except EOFError:
+            return (None, None)
+        except (
+            nme.NMEAMessageError,
+            nme.NMEATypeError,
+            nme.NMEAParseError,
+            nme.NMEAStreamError,
+        ) as err:
+            if self._quitonerror:
+                self._do_error(str(err))
+            parsed_data = str(err)
 
         return (raw_data, parsed_data)
 
+    def _read_bytes(self, size: int) -> bytes:
+        """
+        Read a specified number of bytes from stream.
+
+        :param int size: number of bytes to read
+        :return: bytes
+        :rtype: bytes
+        :raises: EOFError if stream ends prematurely
+        """
+
+        data = self._stream.read(size)
+        if len(data) < size:  # EOF
+            raise EOFError()
+        return data
+
+    def _read_line(self) -> bytes:
+        """
+        Read until end of line (CRLF).
+
+        :return: bytes
+        :rtype: bytes
+        :raises: EOFError if stream ends prematurely
+        """
+
+        data = self._stream.readline()  # NMEA protocol is CRLF terminated
+        if data[-2:] != b"\x0d\x0a":  # EOF
+            raise EOFError()
+        return data
+
+    def _do_error(self, err: str):
+        """
+        Handle error.
+
+        :param str err: error message
+        :raises: UBXParseError if quitonerror = 2
+        """
+
+        if self._quitonerror == ERR_RAISE:
+            raise nme.NMEAParseError(err)
+        if self._quitonerror == ERR_LOG:
+            # pass to error handler if there is one
+            if self._errorhandler is None:
+                print(err)
+            else:
+                self._errorhandler(err)
+
     def iterate(self, **kwargs) -> tuple:
         """
+        DEPRECATED - WILL BE REMOVED IN VERSION >=1.2.21
+        USE STANDARD ITERATOR INSTEAD
         Invoke the iterator within an exception handling framework.
 
-        :param bool quitonerror: (kwarg) Quit on NMEA error True/False (True)
-        :param object errorhandler: (kwarg) Optional error handler (None)
         :return: tuple of (raw_data as bytes, parsed_data as NMEAMessage)
         :rtype: tuple
         :raises: NMEA...Error (if quitonerror is True and stream is invalid)
-
         """
-
-        quitonerror = kwargs.get("quitonerror", True)
-        errorhandler = kwargs.get("errorhandler", None)
 
         while True:
             try:
                 yield next(self)
             except StopIteration:
                 break
-            except (
-                nme.NMEAMessageError,
-                nme.NMEATypeError,
-                nme.NMEAParseError,
-                nme.NMEAStreamError,
-            ) as err:
-                if quitonerror:
-                    raise err
-                if errorhandler is None:
-                    print(err)
-                else:
-                    errorhandler(err)
-                continue
 
     @staticmethod
     def parse(message: bytes, **kwargs) -> object: