+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # pynmeagps Release Notes
 
+### RELEASE 1.2.20rc0
+
+CHANGES:
+
+1. `quitonerror` and `errorhandler` kwargs added to NMEAReader constructor - see Sphinx documentation for details
+2. `NMEAReader.iterate()` method deprecated - use the standard iterator instead e.g. `nmr = NMEAReader(**wkargs): for (raw,parse) in nmr: ...`, passing any `quitonerror` or `errorhandler` kwargs to the NMEAReader constructor.
+
 ### RELEASE 1.0.19
 
 FIXES: